alexa: move intent parsing code to a separate file
It will make it easier to test.

Also, switch to typecheck the program code, as that's necessary
to convert to NN.
